requirements.txt
# Qubes - Sovereign AI Agents
# Requirements - Updated October 2025

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================
pydantic>=2.12.3         # Data validation (updated from 2.5.0)
pyyaml>=6.0.2            # YAML config parsing (updated from 6.0)
python-dotenv>=1.0.1     # Environment variables (updated from 1.0.0)

# ============================================================================
# CRYPTOGRAPHY
# ============================================================================
cryptography>=44.0.0     # AES-256-GCM, PBKDF2, Fernet (updated from 41.0.0)
ecdsa>=0.19.0            # ECDSA secp256k1 signatures (updated from 0.18.0)

# ============================================================================
# STORAGE
# ============================================================================
# Storage is now JSON-based (individual .json files per block)
# LMDB and msgpack removed in favor of native JSON serialization

# ============================================================================
# LOGGING & OBSERVABILITY
# ============================================================================
structlog>=25.4.0                  # Structured logging (updated from 24.1.0)
opentelemetry-api>=1.29.0          # Distributed tracing API (updated from 1.22.0)
opentelemetry-sdk>=1.29.0          # Distributed tracing SDK (updated from 1.22.0)
prometheus-client>=0.21.1          # Prometheus metrics (updated from 0.19.0)

# ============================================================================
# ERROR HANDLING & RESILIENCE
# ============================================================================
tenacity>=9.0.0          # Retry decorators with exponential backoff (updated from 8.2.3)
pybreaker>=1.2.0         # Circuit breaker pattern (updated from 1.0.1)

# ============================================================================
# SYSTEM MONITORING
# ============================================================================
psutil>=6.1.1            # System metrics (CPU, memory, disk) (updated from 5.9.6)

# ============================================================================
# TESTING
# ============================================================================
pytest>=8.3.4            # Testing framework (updated from 7.4.3)
pytest-asyncio>=0.25.2   # Async test support (updated from 0.21.1)
pytest-cov>=6.0.0        # Code coverage (updated from 4.1.0)

# ============================================================================
# SECURITY SCANNING
# ============================================================================
bandit>=1.8.0            # Python security linter (updated from 1.7.5)

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
black>=24.10.0           # Code formatting (updated from 23.12.0)
isort>=5.13.2            # Import sorting (updated from 5.13.0)
mypy>=1.13.0             # Type checking (updated from 1.7.0)
ruff>=0.8.4              # Fast Python linter (updated from 0.1.8)

# ============================================================================
# AI/LLM INTEGRATION (Phase 2)
# ============================================================================
openai>=2.5.0                      # GPT-4, GPT-4o, o1, o3, DALL-E 3 (updated from 1.6.0)
anthropic>=0.71.0                  # Claude 3.5, Claude 4 (updated from 0.8.0)
google-genai>=1.46.0               # Gemini 2.0, 2.5 - NEW SDK (replaces google-generativeai)

# Semantic Search (Phase 2)
faiss-cpu>=1.9.0.post1             # Vector similarity search (updated from 1.7.4)
sentence-transformers>=3.3.1       # Text embeddings for semantic search (updated from 2.2.2)

# ============================================================================
# P2P NETWORKING (Phase 3)
# ============================================================================
# Option A: py-libp2p (native Python, less stable)
# libp2p>=0.2.0           # P2P networking stack (IPFS/Filecoin)
# multiaddr>=0.0.9        # Multiaddr format for P2P addresses

# Option B: libp2p-daemon bridge (recommended, stable)
p2pclient>=0.1.1         # Python bindings for go-libp2p-daemon
multiaddr>=0.0.9         # Multiaddr format for P2P addresses

# Required: Install go-libp2p-daemon binary separately
# go install github.com/libp2p/go-libp2p-daemon/p2pd@latest
# Binary will be at ~/go/bin/p2pd (add to PATH)

# ============================================================================
# BLOCKCHAIN (Phase 4)
# ============================================================================
bitcash>=1.1.0           # Bitcoin Cash with CashTokens support
ipfshttpclient>=0.8.0a2  # IPFS uploads for BCMR metadata (updated from 0.7.0)
aiohttp>=3.11.11         # Async HTTP for Chaingraph GraphQL queries (updated from 3.8.0)

# Health Checks & Orchestrator (Phase 1)
fastapi>=0.119.1         # Health check endpoints (updated from 0.115.0)
uvicorn>=0.34.0          # ASGI server for health API (updated from 0.32.0)

# ============================================================================
# AUDIO INTEGRATION (Phase 7 - Week 19)
# ============================================================================
# Audio I/O
sounddevice>=0.5.1          # Cross-platform audio playback (updated from 0.4.6)
pyaudio>=0.2.14             # Microphone recording
pydub>=0.25.1               # Audio format conversion

# TTS/STT Providers
# openai already included above
elevenlabs>=1.99.3          # ElevenLabs TTS (premium voices) (updated from 0.2.27)
google-cloud-texttospeech>=2.19.1  # Google Cloud TTS (380+ voices) (updated from 2.14.0)
deepgram-sdk>=3.9.3         # DeepGram STT (optional) (updated from 3.0.0)

# Voice Activity Detection
webrtcvad>=2.0.10           # WebRTC VAD for smart recording

# Local Models (Sovereign Mode)
# piper (60MB) - TTS binary, installed separately
# whisper.cpp (140MB) - STT binary, installed separately
# See docs/27_Audio_TTS_STT_Integration.md Section 15 for installation

# ============================================================================
# CLI & UI (Phase 8)
# ============================================================================
typer>=0.15.1               # Modern CLI framework (updated from 0.9.0)
rich>=13.9.4                # Beautiful terminal output (updated from 13.7.0)
click>=8.1.8                # CLI utilities (typer dependency) (updated from 8.1.7)

# ============================================================================
# REAL-TIME UPDATES (Phase 5 - Relationships)
# ============================================================================
websockets>=14.1            # WebSocket server for real-time relationship updates (optional)

# ============================================================================
# WEB BROWSING
# ============================================================================
httpx>=0.27.0               # Async HTTP client for URL fetching
beautifulsoup4>=4.12.3      # HTML parsing and text extraction
